minor minor refactoring
